backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
google-generativeai==0.3.2
supabase==1.2.0
cryptography==41.0.7
python-dotenv==1.0.0
pytest==7.4.3
